Add Documentation Assembler GitHub Action
Introduce a new GitHub Action named "Documentation Assembler" to run specified assembler commands using a Docker image.
